+++ gloss-gk/src-gloss/sbrk.c
@@ -5,7 +5,7 @@ void *_sbrk(int nbytes)
 {
 	void *ret;
 	int _errno;
-	__syscall(__syscall_sbrk, &ret, (void*)nbytes, &_errno);
+	__syscall(__syscall_sbrk, &ret, (void*)(intptr_t)nbytes, &_errno);
 	if(ret == (void*)-1)
 		errno = _errno;
 	return ret;
